+++ kernel/net/netfilter/ipset/ip_set_hash_ipportnet.c
@@ -247,7 +247,7 @@ hash_ipportnet4_uadt(struct ip_set *set, struct nlattr *tb[],
 		ip_set_mask_from_to(ip, ip_to, cidr);
 	}
 
-	if ((ip_to - ip + 1)*(port_to - port + 1) > IPSET_MAX_RANGE)
+	if (((u64)ip_to - ip + 1)*(port_to - port + 1) > IPSET_MAX_RANGE)
 		return -ERANGE;
 
 	port_to = port = ntohs(e.port);
